restructure worker handler execution for more intuitive operator chain
