+++ README.md
@@ -49,10 +49,10 @@ type Interface interface {
 
 Kayron is a change management controller implementing the [operator pattern], in
 this particular case without the involvement of [Kubernetes]. The main goroutine
-for the operator's reconciliation loop is the operator worker handler in
-`pkg/worker/handler/operator/`. Secondary worker handlers are executed within
-their own isolated failure domain, feeding into the primary business logic of
-the operator using concurrency safe cache channels.
+for the operator's reconciliation loop is the operator worker handler running a
+sequence of steps according to their operator functions located in
+`pkg/operator/`. Secondary worker handlers may be executed within their own
+isolated failure domain.
 
 Operators try to continuously drive the current state of a system towards the
 desired state of a system. In our case, the current state is represented by the
