+++ go.mod
@@ -4,7 +4,7 @@ go 1.24
 
 require (
 	github.com/0xSplits/otelgo v0.1.0
-	github.com/0xSplits/workit v0.4.0
+	github.com/0xSplits/workit v0.4.1
 	github.com/aws/aws-sdk-go-v2 v1.37.2
 	github.com/aws/aws-sdk-go-v2/config v1.30.3
 	github.com/aws/aws-sdk-go-v2/service/ecs v1.62.0
@@ -21,7 +21,7 @@ require (
 	github.com/rs/cors v1.11.1
 	github.com/spf13/afero v1.14.0
 	github.com/spf13/cobra v1.9.1
-	github.com/xh3b4sd/choreo v0.3.0
+	github.com/xh3b4sd/choreo v0.4.0
 	github.com/xh3b4sd/logger v0.11.0
 	github.com/xh3b4sd/tracer v1.0.0
 	go.opentelemetry.io/otel/metric v1.37.0
