+++ pkg/operator/cloudformation/cloudformation.go
