+++ pkg/operator/cloudformation/ensure.go
