+++ pkg/operator/cloudformation/registry.go
