+++ pkg/operator/container/detail.go
