+++ pkg/operator/container/ensure.go
