+++ pkg/operator/container/image.go
