+++ pkg/operator/container/image_test.go
