+++ pkg/operator/infrastructure/ensure.go
