+++ pkg/operator/policy/ensure.go
