+++ pkg/operator/reference/ensure.go
