+++ pkg/operator/registry/ensure.go
