+++ pkg/operator/registry/error.go
