+++ pkg/operator/registry/registry.go
