+++ pkg/operator/release/ensure.go
@@ -3,10 +3,10 @@ package release
 import (
 	"fmt"
 
+	"github.com/0xSplits/kayron/pkg/operator/release/resolver"
 	"github.com/0xSplits/kayron/pkg/release/loader"
 	"github.com/0xSplits/kayron/pkg/release/schema"
 	"github.com/0xSplits/kayron/pkg/roghfs"
-	"github.com/0xSplits/kayron/pkg/worker/handler/operator/release/resolver"
 	"github.com/spf13/afero"
 	"github.com/xh3b4sd/tracer"
 )
