+++ pkg/operator/release/resolver/commit.go
