+++ pkg/operator/release/resolver/interface.go
