--- pkg/worker/handler/operator/cooler.go
@@ -1,9 +1,0 @@
-package operator
-
-import (
-	"time"
-)
-
-func (h *Handler) Cooler() time.Duration {
-	return 10 * time.Second
-}
